Ensure packageInfo from networks doesn't clobber crypto
