+++ packages/util-crypto/src/networks.ts
@@ -1,4 +1,4 @@
 // Copyright 2017-2021 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export * from '@polkadot/networks';
+export * from '@polkadot/networks/interfaces';
